refactor(stm): move the IVC proof system to a suffixed module name
