+++ mithril-stm/src/proof_system/halo2_ivc_snark/errors.rs
